Do not run utf-8 with py2
