+++ tests/test_hooks/conan-center/test_conan-center.py
@@ -2,7 +2,6 @@ import os
 import platform
 import textwrap
 import pytest
-import six
 
 from conans import tools
 from conans.client.command import ERROR_INVALID_CONFIGURATION, SUCCESS
@@ -669,23 +668,6 @@ class ConanCenterTests(ConanClientTestCase):
             tools.save('conanfile.py', content=conanfile.replace("{}", it))
             output = self.conan(['create', '.', 'name/version@user/test'])
             self.assertIn("[NO TARGET NAME (KB-H040)] OK", output)
-
-    @pytest.mark.skipif(six.PY2, reason="Python 2 doesn't support UTF-8 by default")
-    def test_non_ascii_characters(self):
-        conanfile = textwrap.dedent("""\
-        from conans import ConanFile
-        class AConan(ConanFile):
-            {}
-            pass
-        """)
-        tools.save('conanfile.py', content=conanfile.replace("{}", "# Conan, the barbarian"))
-        output = self.conan(['create', '.', 'name/version@user/test'])
-        self.assertIn("[NO ASCII CHARACTERS (KB-H047)] OK", output)
-
-        tools.save('conanfile.py', content=conanfile.replace("{}", "# Conan, o bárbaro"))
-        output = self.conan(['create', '.', 'name/version@user/test'])
-        self.assertIn("ERROR: [NO ASCII CHARACTERS (KB-H047)] The line (3) contains a non-ascii character." \
-                      " Only ASCII characters are allowed, please remove it.", output)
 
     def test_delete_option(self):
         conanfile = textwrap.dedent("""\
